refactor(dashboard): extract achievements into useAchievements hook
Move the ~320-line milestone/achievement definition + earned-count memo
verbatim into _hooks/use-achievements.ts. The hook owns its own
useTranslations("Dashboard.body") (matching the page's tb) and takes
only the derived data (recommendations, libraryItems, stats, ratedCount,
streak). Returns the same { list, earned } shape; all consuming effects
and JSX unchanged. Behavior-identical; page.tsx 2328->2014 LOC.
